+++ NEWS.md
@@ -1,3 +1,5 @@
+* Added option to `zero_fill` to filter to specified species, along with more tests to ensure there are zeros to fill. 
+
 # LTMRdata 0.5.0
 
 * In Suisun data, added counts with no corresponding length measurements, e.g., fish within a certain size group were counted and not measured and no fish within that size group were measured. Despite the use of the word "fish" the vast majority of these instances are for invertebrates. Only 334 rows out of 24,289 corresponded to fish.